+++ plugin-code/src/main/java/com/trivago/cluecumber/rendering/pages/renderering/AllFeaturesPageRenderer.java
@@ -21,25 +21,16 @@ import com.trivago.cluecumber.constants.Status;
 import com.trivago.cluecumber.exceptions.CluecumberPluginException;
 import com.trivago.cluecumber.rendering.pages.charts.ChartJsonConverter;
 import com.trivago.cluecumber.rendering.pages.charts.StackedBarChartBuilder;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.Axis;
 import com.trivago.cluecumber.rendering.pages.charts.pojos.Chart;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.Data;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.Dataset;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.Options;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.ScaleLabel;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.Scales;
-import com.trivago.cluecumber.rendering.pages.charts.pojos.Ticks;
 import com.trivago.cluecumber.rendering.pages.pojos.Feature;
-import com.trivago.cluecumber.rendering.pages.pojos.ResultCount;
 import com.trivago.cluecumber.rendering.pages.pojos.pagecollections.AllFeaturesPageCollection;
 import freemarker.template.Template;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
+import java.util.stream.Collectors;
 
 @Singleton
 public class AllFeaturesPageRenderer extends PageRenderer {
@@ -64,85 +55,31 @@ public class AllFeaturesPageRenderer extends PageRenderer {
     }
 
     private void addChartJsonToReportDetails(final AllFeaturesPageCollection allFeaturesPageCollection) {
-        StackedBarChartBuilder stackedBarChartBuilder = new StackedBarChartBuilder(chartConfiguration);
-        Chart chart = stackedBarChartBuilder.build();
-
-        Data data = new Data();
-        chart.setData(data);
-
-        List<Dataset> datasets = new ArrayList<>();
-
         List<Integer> passed = new ArrayList<>();
         List<Integer> failed = new ArrayList<>();
         List<Integer> skipped = new ArrayList<>();
 
-        int maxY = 0;
-        for (Map.Entry<Feature, ResultCount> entry : allFeaturesPageCollection.getFeatureResultCounts().entrySet()) {
-            passed.add(entry.getValue().getPassed());
-            failed.add(entry.getValue().getFailed());
-            skipped.add(entry.getValue().getSkipped());
-            maxY = entry.getValue().getTotal();
-        }
+        allFeaturesPageCollection.getFeatureResultCounts().forEach((key, value) -> {
+            passed.add(value.getPassed());
+            failed.add(value.getFailed());
+            skipped.add(value.getSkipped());
+        });
 
-        Dataset passedDataset = new Dataset();
-        passedDataset.setLabel(Status.PASSED.getStatusString());
-        passedDataset.setData(passed);
-        List<String> passedBG = new ArrayList<>(Collections.nCopies(passed.size(), chartConfiguration.getPassedColorRgbaString()));
-        passedDataset.setBackgroundColor(passedBG);
-        datasets.add(passedDataset);
+        List<String> keys = allFeaturesPageCollection.getFeatureResultCounts()
+                .keySet()
+                .stream()
+                .map(Feature::getName)
+                .collect(Collectors.toList());
 
-        Dataset failedDataset = new Dataset();
-        failedDataset.setLabel(Status.FAILED.getStatusString());
-        failedDataset.setData(failed);
-        List<String> failedBG = new ArrayList<>(Collections.nCopies(failed.size(), chartConfiguration.getFailedColorRgbaString()));
-        failedDataset.setBackgroundColor(failedBG);
-        datasets.add(failedDataset);
-
-        Dataset skippedDataset = new Dataset();
-        skippedDataset.setLabel(Status.SKIPPED.getStatusString());
-        skippedDataset.setData(skipped);
-        List<String> skippedBG = new ArrayList<>(Collections.nCopies(skipped.size(), chartConfiguration.getSkippedColorRgbaString()));
-        skippedDataset.setBackgroundColor(skippedBG);
-        datasets.add(skippedDataset);
-
-        data.setDatasets(datasets);
-
-        List<String> keys = new ArrayList<>();
-        for (Feature feature : allFeaturesPageCollection.getFeatureResultCounts().keySet()) {
-            keys.add(feature.getName());
-        }
-        data.setLabels(keys);
-
-        Options options = new Options();
-        Scales scales = new Scales();
-        List<Axis> xAxes = new ArrayList<>();
-        Axis xAxis = new Axis();
-        xAxis.setStacked(true);
-        Ticks xTicks = new Ticks();
-        xTicks.setDisplay(false);
-        xAxis.setTicks(xTicks);
-        ScaleLabel xScaleLabel = new ScaleLabel();
-        xScaleLabel.setDisplay(true);
-        xScaleLabel.setLabelString(allFeaturesPageCollection.getTotalNumberOfFeatures() + " Feature(s)");
-        xAxis.setScaleLabel(xScaleLabel);
-        xAxes.add(xAxis);
-        scales.setxAxes(xAxes);
-
-        List<Axis> yAxes = new ArrayList<>();
-        Axis yAxis = new Axis();
-        yAxis.setStacked(true);
-        Ticks yTicks = new Ticks();
-        yAxis.setTicks(yTicks);
-        ScaleLabel yScaleLabel = new ScaleLabel();
-        yScaleLabel.setDisplay(true);
-        yScaleLabel.setLabelString("Number of Scenarios");
-        yAxis.setScaleLabel(yScaleLabel);
-        yAxis.setStepSize(maxY);
-        yAxes.add(yAxis);
-        scales.setyAxes(yAxes);
-
-        options.setScales(scales);
-        chart.setOptions(options);
+        Chart chart =
+                new StackedBarChartBuilder(chartConfiguration)
+                        .setLabels(keys)
+                        .setxAxisLabel(allFeaturesPageCollection.getTotalNumberOfFeatures() + " Feature(s)")
+                        .setyAxisLabel("Number of Scenarios")
+                        .addValues(passed, Status.PASSED)
+                        .addValues(failed, Status.FAILED)
+                        .addValues(skipped, Status.SKIPPED)
+                        .build();
 
         allFeaturesPageCollection.getReportDetails().setChartJson(convertChartToJson(chart));
     }